#12 refactors code to remove all references to old request params
Basically just getting rid of all references to "itemType" and "foodType". In some places, references to "request" have been altered to reflect its new params. This changes a lot on the client side and basically strips the app of any of its core functionality. As dramatic as that may sound, this refactoring was a necessary first step towards building our second iteration.